+++ src/main/kotlin/com/abeade/plugin/figma/ui/PreviewPanel.kt
@@ -1,0 +1,11 @@
+package com.abeade.plugin.figma.ui
+
+import com.abeade.plugin.figma.ui.preview.QuickDrawPanel
+import javax.swing.JLabel
+import javax.swing.JPanel
+
+class PreviewPanel {
+    lateinit var quickDrawPanel: QuickDrawPanel
+    lateinit var labelSize: JLabel
+    lateinit var mainPanel: JPanel
+}